requirements.txt
langchain>=0.1.0
langchain-community>=0.0.27
langchain-core>=0.1.32
langchain-openai>=0.0.8
langchain-pinecone>=0.0.1
openai>=1.12.0
pinecone>=3.0.2
python-dotenv>=1.0.1
unstructured[pdf,xlsx,docx]>=0.17.2